Renamed files to distinguish between kmknn and annoy.
