+++ src/init.cpp
@@ -1,4 +1,4 @@
-#include "kmknn.h"
+#include "init.h"
 #include "R_ext/Rdynload.h"
 #include "R_ext/Visibility.h"
 
@@ -7,8 +7,8 @@
 extern "C" {
 
 static const R_CallMethodDef all_call_entries[] = {
-    REGISTER(find_knn, 7),
-    REGISTER(query_knn, 8),
+    REGISTER(find_kmknn, 7),
+    REGISTER(query_kmknn, 8),
     REGISTER(find_neighbors, 7),
     REGISTER(query_neighbors, 8),
