+++ src/init.h
@@ -1,0 +1,22 @@
+#ifndef INIT_H
+#define INIT_H
+
+#include "Rcpp.h"
+
+extern "C" {
+
+SEXP find_kmknn(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP); 
+
+SEXP query_kmknn(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP); 
+
+SEXP find_neighbors(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP); 
+
+SEXP query_neighbors(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP); 
+
+SEXP build_annoy(SEXP, SEXP, SEXP);
+
+SEXP find_annoy(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
+
+}
+
+#endif
