+++ src/kmknn.cpp
@@ -1,4 +1,4 @@
-#include "objects.h"
+#include "kmknn.h"
 #include "utils.h"
 #define USE_UPPER 0
